Add snackbar error freedback for out of bounds click (inactive routes)
