+++ client/src/components/CollisionModal.jsx
@@ -1,4 +1,4 @@
-import { useState, memo, useContext } from "react";
+import { useState, useContext } from "react";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Dialog from "@mui/material/Dialog";
@@ -7,23 +7,13 @@ import DialogContent from "@mui/material/DialogContent";
 import DialogContentText from "@mui/material/DialogContentText";
 import DialogTitle from "@mui/material/DialogTitle";
 import FormControl from "@mui/material/FormControl";
-import FormControlLabel from "@mui/material/FormControlLabel";
 import InputLabel from "@mui/material/InputLabel";
 import MenuItem from "@mui/material/MenuItem";
 import Select from "@mui/material/Select";
-import Switch from "@mui/material/Switch";
 import Snackbar from "@mui/material/Snackbar";
 import Alert from "@mui/material/Alert";
 import "../styles/CollisionModal.css";
 import ReverseContext from "../context/ReverseContext";
-import { ThemeProvider, createTheme } from "@mui/material/styles";
-import CssBaseline from "@mui/material/CssBaseline";
-
-const darkTheme = createTheme({
-  palette: {
-    mode: "dark",
-  },
-});
 
 const CollisionModal = ({ open, setOpen, relationships, fieldInfo }) => {
   // console.log("source types: ", relationships);
@@ -68,57 +58,54 @@ const CollisionModal = ({ open, setOpen, relationships, fieldInfo }) => {
 
   return (
     <div className="collision-modal-container">
-      <ThemeProvider theme={darkTheme}>
-        <CssBaseline />
-        <Dialog open={open} onClose={handleClose}>
-          <DialogTitle>Collision Detected</DialogTitle>
-          <DialogContent>
-            <DialogContentText>
-              Multiple active routes lead to this field. Which source would you
-              like to add it to?
-            </DialogContentText>
-            <Box
-              noValidate
-              component="form"
-              sx={{
-                display: "flex",
-                flexDirection: "column",
-                m: "auto",
-                width: "fit-content",
-              }}
-            >
-              <FormControl sx={{ mt: 2, minWidth: 140 }}>
-                <InputLabel htmlFor="source-select">Source</InputLabel>
-                <Select
-                  autoFocus
-                  value={source}
-                  onChange={handleSourceChange}
-                  label="source-select"
-                  inputProps={{
-                    name: "source-select",
-                    id: "source-select",
-                  }}
-                >
-                  {menuItems}
-                </Select>
-              </FormControl>
-            </Box>
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={handleClose}>Select</Button>
-          </DialogActions>
-          <Snackbar
-            open={snackbarOpen}
-            autoHideDuration={2000}
-            onClose={handleSnackbarClose}
-            anchorOrigin={{ vertical: "center", horizontal: "center" }}
+      <Dialog open={open} onClose={handleClose}>
+        <DialogTitle>Collision Detected</DialogTitle>
+        <DialogContent>
+          <DialogContentText>
+            Multiple active routes lead to this field. Which source would you
+            like to add it to?
+          </DialogContentText>
+          <Box
+            noValidate
+            component="form"
+            sx={{
+              display: "flex",
+              flexDirection: "column",
+              m: "auto",
+              width: "fit-content",
+            }}
           >
-            <Alert severity="warning" sx={{ width: "100%" }}>
-              Please select an option
-            </Alert>
-          </Snackbar>
-        </Dialog>
-      </ThemeProvider>
+            <FormControl sx={{ mt: 2, minWidth: 140 }}>
+              <InputLabel htmlFor="source-select">Source</InputLabel>
+              <Select
+                autoFocus
+                value={source}
+                onChange={handleSourceChange}
+                label="source-select"
+                inputProps={{
+                  name: "source-select",
+                  id: "source-select",
+                }}
+              >
+                {menuItems}
+              </Select>
+            </FormControl>
+          </Box>
+        </DialogContent>
+        <DialogActions>
+          <Button onClick={handleClose}>Select</Button>
+        </DialogActions>
+        <Snackbar
+          open={snackbarOpen}
+          autoHideDuration={2000}
+          onClose={handleSnackbarClose}
+          anchorOrigin={{ vertical: "center", horizontal: "center" }}
+        >
+          <Alert severity="warning" sx={{ width: "100%" }}>
+            Please select an option
+          </Alert>
+        </Snackbar>
+      </Dialog>
     </div>
   );
 };
