+++ client/src/main.jsx
@@ -4,12 +4,23 @@ import App from "./App";
 import { ReactFlowProvider } from "reactflow";
 import { ContextProvider } from "./context/ReverseContext";
 import "./styles/main.css";
+import { ThemeProvider, createTheme } from "@mui/material/styles";
+import CssBaseline from "@mui/material/CssBaseline";
+
+const darkTheme = createTheme({
+  palette: {
+    mode: "dark",
+  },
+});
 
 ReactDOM.render(
   <React.StrictMode>
     <ReactFlowProvider>
       <ContextProvider>
-        <App />
+        <ThemeProvider theme={darkTheme}>
+          <CssBaseline />
+          <App />
+        </ThemeProvider>
       </ContextProvider>
     </ReactFlowProvider>
   </React.StrictMode>,
